requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
discord.py==1.7.3
Flask==2.0.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
outcome==1.1.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlitedict==1.7.0
trio==0.19.0
trio-websocket==0.9.2
twilio==7.4.0
typing_extensions==4.0.1
undetected-chromedriver==3.1.3
urllib3==1.26.7
websockets==10.1
Werkzeug==2.0.2
wsproto==1.0.0
yarl==1.7.2
